update create product -fix  validate unique product

+++ product/src/main/java/com/lexuancong/product/viewmodel/product/databinding/BaseProductPropertiesRequire.java
@@ -1,5 +1,6 @@
 package com.lexuancong.product.viewmodel.product.databinding;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 // những thuộc tính cơ ba nhất của sản phẩm bắt buộc phải cs khi thêm sp hay lamf gì đó
@@ -11,8 +12,8 @@ public interface BaseProductPropertiesRequire {
 
     String gtin();
 
-    Double price();
-
+//    Double price();
+    BigDecimal price();
     Long avatarImageId();
 
     List<Long> imageIds();
